@@ -2,13 +2,14 @@ package com.dinner.dinner;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
 import android.widget.CheckBox;
 import android.widget.EditText;
 import android.widget.TextView;
-import android.widget.Toast;
+
+import androidx.appcompat.app.AppCompatActivity;
+
 
 public class LoginActivity extends AppCompatActivity {
 
@@ -20,8 +21,9 @@ public class LoginActivity extends AppCompatActivity {
         final EditText username = findViewById(R.id.username);
         final EditText password = findViewById(R.id.password);
         final CheckBox rememberMe = findViewById(R.id.remember_me);
+
+
         Button loginBtn = findViewById(R.id.login_btn);
-        Button registerBtn = findViewById(R.id.register_btn);
 
         final User user = new User(LoginActivity.this);
 
@@ -30,41 +32,48 @@ public class LoginActivity extends AppCompatActivity {
         if (rememberMe.isChecked()) {
             username.setText(user.getUsernameForLogin(), TextView.BufferType.EDITABLE);
             password.setText(user.getPasswordForLogin(), TextView.BufferType.EDITABLE);
+
         } else {
+
             username.setText("", TextView.BufferType.EDITABLE);
             password.setText("", TextView.BufferType.EDITABLE);
         }
 
+        Button registerBtn = findViewById(R.id.register_btn);
+
         loginBtn.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                //čia rašomas kodas, kuris bus vykdomas ant mygtuko paspaudimo//
+                //Rašomas kodas, kuris bus vykdomas ant mygtuko paspaudimo//
                 String username2 = username.getText().toString();
                 String password2 = password.getText().toString();
-                boolean validUserName = InputValidator.isCredentialsValid(username2);
-                boolean validPassword = InputValidator.isCredentialsValid(password2);
-                // purge error logs
+
+                //Klaidu zurnalo isvalymas//
                 username.setError(null);
                 password.setError(null);
-                if (!validPassword || !validUserName) {
-                    //show error
-                    username.setError(getResources().getString(R.string.login_invalid_credentials_message));
-                    username.requestFocus();
-                }
 
-                //----------------------------------------------------iš kur-------------į kur---------//
-                if (validUserName && validPassword) {
+                if (Validation.isCredentialsValid(username2) && Validation.isCredentialsValid(password2)) {
                     user.setUsernameForLogin(username2);
                     user.setPasswordForLogin(password2);
                     if (rememberMe.isChecked()) {
-                        user.setRememberMeKey(true);
+                        user.setRememberMeKeyForLogin(true);
                     } else {
-                        user.setRememberMeKey(false);
-                    }
 
-                    Toast.makeText(LoginActivity.this, "Welcome back, : " + username2, Toast.LENGTH_SHORT).show();
+                        user.setRememberMeKeyForLogin(false);
+                    }
+                    //----------------------------------------------------iš kur-------------į kur---------//
                     Intent gotoSearchActivity = new Intent(LoginActivity.this, SearchActivity.class);
                     startActivity(gotoSearchActivity);
+                } else {
+
+                    /*Toast.makeText(LoginActivity.this, "Username: " + username2 + "\n" + "Password: "
+                            + password2, Toast.LENGTH_SHORT).show();*/
+                    username.setError(getResources().getString(R.string.login_invalid_credentials_message));
+                    username.requestFocus();
+                    password.setError(getResources().getString(R.string.login_invalid_credentials_message));
+                    password.requestFocus();
+                    /* Toast.makeText(LoginActivity.this,getResources().getString(R.string.login_invalid_credentials_message),
+                          Toast.LENGTH_SHORT).show();*/
                 }
             }
         });
@@ -72,13 +81,15 @@ public class LoginActivity extends AppCompatActivity {
         registerBtn.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
+                //Rašomas kodas, kuris bus vykdomas ant mygtuko paspaudimo, perejimas i register langa//
+
+                //----------------------------------------------------iš kur-------------į kur---------//
                 Intent gotoRegisterActivity = new Intent(LoginActivity.this, RegisterActivity.class);
                 startActivity(gotoRegisterActivity);
             }
         });
-
     }
 
 }
 
-//TODO email regex
+
