@@ -12,19 +12,17 @@ public class User {
     private static final String PREFERENCES_PACKAGE_NAME = "com.dinner.dinner";
     private static final String USERNAME_KEY = "username";
     private static final String PASSWORD_KEY = "password";
-    private static final String REMEMBER_ME_KEY  = "rememberMe";
+    private static final String REMEMBER_ME_KEY = "rememberMe";
 
-    /*
-    constructor for new user registration
-     */
+    // Konstruktorius skirtas naujo vartotojo registracijai
     public User(String username, String password, String email) {
         this.username = username;
         this.password = password;
         this.email = email;
+
     }
-    /*
-    Constructor for user login
-     */
+
+    // Konstruktorius skirtas vartotojo prisijungimui
     public User(Context context) {
         this.sharedPreferences = context.getSharedPreferences(User.PREFERENCES_PACKAGE_NAME,
                 Context.MODE_PRIVATE);
@@ -47,34 +45,40 @@ public class User {
     }
 
     public void setPasswordForRegistration(String password) {
-        this.password = password;
+        this.username = password;
     }
 
     public void setEmailForRegistration(String email) {
         this.email = email;
     }
 
-    public String getUsernameForLogin() {
-        return this.sharedPreferences.getString(USERNAME_KEY, "");
+    public String getUsernameForLogin(){
+        return this.sharedPreferences.getString(USERNAME_KEY,"");
+
+    }
+    public void setUsernameForLogin(String username){
+
+        this.sharedPreferences.edit().putString(USERNAME_KEY, username).commit();
     }
 
-    public void setUsernameForLogin(String username) {
-        this.sharedPreferences.edit().putString(USERNAME_KEY, username).apply();
+    public String getPasswordForLogin(){
+        return this.sharedPreferences.getString(PASSWORD_KEY,"");
+
     }
 
-    public String getPasswordForLogin() {
-        return this.sharedPreferences.getString(PASSWORD_KEY, "");
+    public void setPasswordForLogin(String password){
+
+        this.sharedPreferences.edit().putString(PASSWORD_KEY, password).commit();
     }
 
-    public void setPasswordForLogin(String password) {
-        this.sharedPreferences.edit().putString(PASSWORD_KEY, password).apply();
-    }
-
-    public boolean isRememberedForLogin() {
+    public boolean isRememberedForLogin(){
         return this.sharedPreferences.getBoolean(REMEMBER_ME_KEY, false);
+
     }
 
-    public void setRememberMeKey(boolean rememberMeKey) {
-        this.sharedPreferences.edit().putBoolean(REMEMBER_ME_KEY, rememberMeKey).apply();
+    public void setRememberMeKeyForLogin(boolean rememberMeKey){
+        this.sharedPreferences.edit().putBoolean(REMEMBER_ME_KEY, rememberMeKey).commit();
+
     }
+
 }
